updated code
I removed the collision code from the game_function and moved it to the move function in the Character and Enemy classes. Additionally, every room has its own objects, so I introduced the current_room variable. This way, we don't have to repeatedly modify the collision code for different rooms. I also added an image_path feature, allowing us to change the image of both enemies and rooms/levels.